+++ tests/testthat/test-combine-modeltime-tables.R
@@ -34,13 +34,14 @@ model_tbl_2 <- modeltime_table(model_fit_prophet)
 
 model_tbl_3 <- modeltime_table(model_fit_ets)
 
-
+calib_tbl <- model_tbl_1 %>%
+    modeltime_calibrate(testing(splits))
 
 # Tests ----
 
 test_that("combine_modeltime_table(): succeeds with mdl_time_tbl classes", {
 
-    model_tbl_combo <- combine_model_tables(model_tbl_1, model_tbl_2, model_tbl_3)
+    model_tbl_combo <- combine_modeltime_tables(model_tbl_1, model_tbl_2, model_tbl_3)
 
     expect_s3_class(model_tbl_combo, "mdl_time_tbl")
 
@@ -51,7 +52,7 @@ test_that("combine_modeltime_table(): succeeds with mdl_time_tbl classes", {
 test_that("combine_modeltime_table(): fails with non-mdl_time_tbl classes", {
 
     expect_error({
-        combine_model_tables(
+        combine_modeltime_tables(
             model_tbl_1, model_tbl_2, model_tbl_3,
             # Bad
             model_fit_arima
@@ -60,7 +61,27 @@ test_that("combine_modeltime_table(): fails with non-mdl_time_tbl classes", {
 
 })
 
+test_that("combine_modeltime_table(): Removes columns when combining calibration tbl", {
 
+    expect_message({
+        combine_modeltime_tables(
+            model_tbl_1, model_tbl_2, model_tbl_3,
+            # Calibrated
+            calib_tbl, calib_tbl
+        )
+    })
+
+    combo_tbl <- combine_modeltime_tables(
+        model_tbl_1, model_tbl_2, model_tbl_3,
+        # Calibrated
+        calib_tbl
+    )
+
+    expect_equal(ncol(combo_tbl), 3)
+
+    expect_equal(nrow(combo_tbl), 4)
+
+})